The constructor now checks for, ignores, and warns about any non-supported name value pairs. Does this using simple string comparison between argument names and uitabgroup properties.
This should allow old code to run without removing the arguments for old features and allow them to be reinstated automatically with updates to uitab.

Note: It does this for any which are not properties of uitab, so it will also warn instead of error if typos are present in the argument names.