Added the input warning and ignoring to adding uitabs to match the same behaviour as when the tabgroup is created.
In both instances, the script now takes all non-matching argument names and attempts validatestring before ignoring them. Again the error has been demoted to a warn when the validation fails.

As this may cause silent errors, when a more comprehensive list of removed features is available it may be worth coding in a line which decides to warn (removed feature) or error (feature could not be validated and no matching feature name ever existed).